+++ src/Ergosfare.Contracts/Commands/ICommandPostInterceptor[TCommand].cs
@@ -1,0 +1,5 @@
+using Ergosfare.Core.Abstractions.Handlers;
+
+namespace Ergosfare.Contracts;
+
+public interface ICommandPostInterceptor<in TCommand>: ICommand, IAsyncPostInterceptor<TCommand>  where TCommand : ICommand;