+++ src/Ergosfare.Contracts/Events/IEventPostInterceptor.cs
@@ -1,0 +1,5 @@
+using Ergosfare.Core.Abstractions.Handlers;
+
+namespace Ergosfare.Contracts;
+
+public interface IEventPostInterceptor:  IEvent, IAsyncPostInterceptor<IEvent>;