+++ src/Ergosfare.Contracts/Events/IEventPostInterceptor[TEvent].cs
@@ -1,0 +1,5 @@
+using Ergosfare.Core.Abstractions.Handlers;
+
+namespace Ergosfare.Contracts;
+
+public interface IEventPostInterceptor<in TEvent> : IEvent, IAsyncPostInterceptor<TEvent> where TEvent : notnull;