+++ src/Ergosfare.Core/Properties/AssemblyInfo.cs
@@ -1,5 +1,6 @@
 ﻿using System.Runtime.CompilerServices;
 
+[assembly:InternalsVisibleTo("Ergosfare.Events.Test")]
 [assembly:InternalsVisibleTo("Ergosfare.Core.Test")]
 [assembly:InternalsVisibleTo("Ergosfare.Core.Extensions.MicrosoftDependencyInjection.Test")]
 [assembly:InternalsVisibleTo("Ergosfare.Test")]