+++ CHANGELOG.md
@@ -1,6 +1,46 @@
-## 🌟 v0.0.5e 
+# 🌟 v0.0.6e 🛠️ Event, Command & Query module, unit tests & code coverage
 
-### 🛠️ CommandMocule unit tests & code coverage
+ 
+#### 📊 New Features: Event module
+* **🔹IEventExceptionInterceptor:** interface added, Event module now supports non generic ExceptionInterceptors.
+* **🔹IEventExceptionInterceptor\<TEvent\>:** interface added, Event module now supports generic`<TEvent>` ExceptionInterceptors.
+* **🔹IEventPreInterceptor:** interface added, Event module now supports non generic PreInterceptors.
+* **🔹IEventPreInterceptor\<TEvent\> :** interface added, Event module now supports generic`<TEvent>` PreInterceptors.
+* **🔹IEventPostInterceptor :** interface added, Event module now supports non generic PostInterceptors.
+* **🔹IEventPostInterceptor\<TEvent\> :** interface added, Event module now supports generic`<TEvent>` PostInterceptors.
+
+
+#### 📊 New Features: Command module
+* **🔹ICommandExceptionInterceptor:** interface added, command module now supports non generic ExceptionInterceptors.
+* **🔹ICommandExceptionInterceptor\<TEvent\>:** interface added, command module now supports generic`<TEvent>` ExceptionInterceptors.
+* **🔹ICommandPreInterceptor:** interface added, command module now supports non generic PreInterceptors.
+* **🔹ICommandPreInterceptor\<TEvent\> :** interface added, command module now supports generic`<TEvent>` PreInterceptors.
+* **🔹ICommandPostInterceptor :** interface added, Command module now supports non generic PostInterceptors.
+* **🔹ICommandPostInterceptor\<TEvent\> :** interface added, Command module now supports generic`<TEvent>` PostInterceptors.
+* **🔹ICommandPostInterceptor\<TEvent,TResult\> :** interface added, Command module now supports generic`<TEvent, TResult>` PostInterceptors.
+
+
+#### 📊 New Features: Query module
+* **🔹IQueryExceptionInterceptor:** interface added, command module now supports non generic ExceptionInterceptors.
+* **🔹IQueryExceptionInterceptor\<TQuery\>:** interface added, query module now supports generic`<TQuery>` ExceptionInterceptors.
+* **🔹IQueryPreInterceptor:** interface added, query module now supports non generic PreInterceptors.
+* **🔹IQueryPreInterceptor\<TQuery\> :** interface added, query module now supports generic`<TQuery>` PreInterceptors.
+* **🔹IQueryPostInterceptor :** interface added, Query module now supports non generic PostInterceptors.
+* **🔹IQueryPostInterceptor\<TQuery\> :** interface added, Query module now supports generic`<TQuery>` PostInterceptors.
+* **🔹IQueryPostInterceptor\<TQuery,TResult\> :** interface added, query module now supports generic`<TQuery, TResult>` PostInterceptors.
+
+### ✅ Test Coverage Milestone
+
+* **💯 100% test coverage** for `Ergosfare.Events`, `Ergosfare.Events.Abstractions`, `Ergosfare.Events.Extensions.MicrosfotDependencyInjection`.
+* **💯 100% test coverage** for `Ergosfare.Queries`, `Ergosfare.Queries.Abstractions`, `Ergosfare.Queries.Extensions.MicrosfotDependencyInjection`.
+
+* **📊 Total project coverage:** 99%.
+
+___
+
+# 🌟 v0.0.5e 
+
+### 🛠️ CommandModule unit tests & code coverage
 
 #### No breaking changes
 * **🔹Refactor:** CommandModuleBuilder.Register<T>() now internally calls CommandModuleBuilder.Register(Type T).
@@ -16,7 +56,7 @@
 
 
 
-## 🌟 v0.0.4e – Core Contracts Refactor & Exception Interceptor
+# 🌟 v0.0.4e – Core Contracts Refactor & Exception Interceptor
 
 ### 🛠️ Core Enhancements
 
@@ -31,7 +71,7 @@
 
 
 
-## v0.0.3e
+# 🌟 v0.0.3e
 
 ### General